added initial support for Windows capture with Aperture.CLI
